+++ sound-mind-studio/tests/test_about_dialog.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <QObject>
+
+class AboutDialogTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void showsTheAppNameAndVersion();
+    void hasACloseButtonThatClosesTheDialog();
+};